more manual testing

+++ tests/test_logging.py
@@ -4,28 +4,62 @@ import pandas as pd
 import limnopapers.utils as utils
 import limnopapers.limnopapers as lp
 
-# read test data from json
-test_data = utils.load_dict_from_file("tests/test_data.json")
 
-# ----
+def test_q_mark_logging():
 
-# pass to get_posts_
-posts = []
-posts_raw = lp.get_posts_("test", feed_dict=test_data)
-posts.append(posts_raw)
+    # read test data from json
+    test_data_raw = utils.load_dict_from_file("tests/test_data.json")
 
-os.remove("tests/test_log.csv")
+    # ---- ensure logging handles middle q marks ----
+    test_data = test_data_raw.copy()
+    test_data["entries"] = [test_data["entries"][0]]
 
-data = lp.get_papers(posts=posts, log_path="tests/test_log.csv")
+    posts = []
+    posts_raw = lp.get_posts_("test", feed_dict=test_data)
+    posts.append(posts_raw)
 
-lp.limnotoots(
-    tweet=False,
-    interactive=True,
-    data=data,
-    log_path="tests/test_log.csv",
-    ignore_all=True,
-)
+    if os.path.exists("tests/test_log.csv"):
+        os.remove("tests/test_log.csv")
 
-log = pd.read_csv("tests/test_log.csv").iloc[-1:, :]
+    data = lp.get_papers(posts=posts, log_path="tests/test_log.csv")
 
-log
+    tooted = lp.limnotoots(
+        tweet=False,
+        interactive=True,
+        data=data,
+        log_path="tests/test_log.csv",
+        ignore_all=True,
+    )
+
+    log = pd.read_csv("tests/test_log.csv").iloc[1:, :].reset_index(drop=True)
+
+    assert utils.has_q_mark(log["title"][0])
+
+    # ---- ensure no toots in log are re-tooted ----
+    test_data = test_data_raw.copy()
+
+    posts = []
+    posts_raw = lp.get_posts_("test", feed_dict=test_data)
+    posts.append(posts_raw)
+
+    data = lp.get_papers(posts=posts, log_path="tests/test_log.csv")
+    tooted = lp.limnotoots(
+        tweet=False,
+        interactive=True,
+        data=data,
+        log_path="tests/test_log.csv",
+        ignore_all=True,
+    )
+    # pd.read_csv("tests/test_log.csv").iloc[1:, :].reset_index(drop=True)
+    assert len(tooted) == 2
+
+    data = lp.get_papers(posts=posts, log_path="tests/test_log.csv")
+    tooted = lp.limnotoots(
+        tweet=False,
+        interactive=True,
+        data=data,
+        log_path="tests/test_log.csv",
+        ignore_all=True,
+    )
+
+    assert len(tooted) == 0
